Reduced socket ping interval
Fixing an issue with Heroku's websockets. If WS client is inactive for 55 seconds, it is disconnected. Previously, the pinging interval was of 30 seconds. So, if just one packet is lost on the way to server, connection will be lost, too. Now, the pinging interval is set to 20 seconds, giving a time for client to ping one more time.